Add style to body and accordion components to look like mockup
Wrap accordion body elements in a div for a consistent border. Wrap HTML body in a container div to have a max width for content. Body is now a flexbox to center the container div. Accordion elements are given margin, padding, borders, and background in order to resemble the mockup.